V6 rename themes (#956)
* Rename themes

* update file refs

* change legacy back to v5
