+++ src/theme/index.js
@@ -1,4 +1,4 @@
-export { default as defaultTheme } from './src/themes/v6'
+export { default as defaultTheme } from './src/themes/default'
 export {
   default as ThemeContext,
   ThemeProvider,
