+++ src/theme/src/themes/default/component-specific/avatarColors.js
