+++ src/theme/src/themes/default/component-specific/badgeColors.js
