+++ src/theme/src/themes/default/component-specific/getAvatarInitialsFontSize.js
