+++ src/theme/src/themes/default/component-specific/getAvatarProps.js
