+++ src/theme/src/themes/default/component-specific/getSwitchClassName.js
