+++ src/theme/src/themes/default/component-specific/spinnerColor.js
