+++ src/theme/stories/index.stories.js
@@ -5,7 +5,7 @@ import Component from '@reactions/component'
 import { Heading } from '../../typography'
 import { Checkbox } from '../../checkbox'
 import v5Theme from '../src/themes/v5'
-import v6Theme from '../src/themes/v6'
+import v6Theme from '../src/themes/default'
 import { TextInput } from '../../text-input'
 import { Textarea } from '../../textarea'
 import { Button } from '../../buttons'
